EE10.05: Extract platform-adapters.ts; move naming and artifact helpers
- New tools/delivery/platform-adapters.ts: all ~20 platform adapter wrappers
  that inject _config.runtime (and packageManager) into platform calls
- tools/delivery/planning.ts: add relativeToRepo (exported), BranchMatch type,
  deriveBranchName, deriveWorktreePath, preferDeliveryBranch (private),
  findExistingBranch
- tools/delivery/ticket-flow.ts: add materializeTicketContext and its private
  helpers (ticketHandoffFileName, copyFileIntoWorktree, copyTicketScopedArtifacts)
- tools/delivery/orchestrator.ts: remove all moved definitions; import from new
  modules; re-export public surface unchanged
- New test/planning.test.ts: deriveBranchName, deriveWorktreePath, findExistingBranch
- New test/ticket-flow.test.ts: materializeTicketContext (verbatim migration)
- test/orchestrator.test.ts: remove migrated describe blocks, clean up imports

Zero behavior change. bun test: 360 pass.
