+++ tools/delivery/test/planning.test.ts
@@ -1,0 +1,40 @@
+import { describe, expect, it } from 'bun:test';
+
+import {
+  deriveBranchName,
+  deriveWorktreePath,
+  findExistingBranch,
+} from '../planning';
+
+describe('planning', () => {
+  it('derives deterministic branch and worktree names', () => {
+    expect(
+      deriveBranchName({
+        id: 'P2.03',
+        slug: 'readme-and-real-world-config-example',
+      }),
+    ).toBe('agents/p2-03-readme-and-real-world-config-example');
+    expect(deriveWorktreePath('/tmp/pirate_claw', 'P2.03')).toBe(
+      '/tmp/pirate_claw_p2_03',
+    );
+  });
+
+  it('prefers existing ticket-id branch matches over title-derived names', () => {
+    expect(
+      findExistingBranch(
+        [
+          'agents/p2-02-movie-matcher-missing-codec',
+          'agents/p2-03-readme-config-live-verification',
+          'agents/p2-04-rename-cli-config',
+        ],
+        {
+          id: 'P2.02',
+          slug: 'movie-matcher-allows-missing-codec',
+        },
+      ),
+    ).toEqual({
+      branch: 'agents/p2-02-movie-matcher-missing-codec',
+      source: 'ticket-id',
+    });
+  });
+});
